refactor(cli): flatten user commands into command package
